+++ src/main/java/ru/yandex/practicum/filmorate/model/User.java
@@ -6,7 +6,7 @@ import java.time.LocalDate;
 
 @Data
 public class User {
-    private int id; // идентификатор
+    private long id; // идентификатор
     private String email; // электронная почта
     private String login; // логин пользователя
     private String name; // имя для отображения
